client: Remove autoclient & bump to v3
Adding the autoclient to vspd created a circular dependency between vspd
and dcrwallet which is cumbersome and not worth the maintence burden.

At this point there are no known consumers of the vspd version of
autoclient, so removing it is not expected to cause problems for
anybody.

The autoclient will continue to live in the dcrwallet repo.