webapi: Characterization test gin JSON binding.
Tests to determine exactly how gin decodes JSON into go maps. It is non
obvious because go and JSON values do not map 1-to-1 - a JSON map can be
empty, null, or it can be missing completely, whereas in go it can only
be nil or empty.

Having these tests ensures that the behaviour does not change
unexpectedly when a new version of gin is pulled.